+++ trainctl/uitrack/uitrack.h
@@ -13,7 +13,7 @@
 
 void uitrack_run_tick(uint32_t notif_flags, uint32_t tick, uint32_t dt);
 
-void impl_uitrack_change_blk(int blk, int v, int trn, int sblk);
+void impl_uitrack_change_blk(uint8_t blk, uint8_t v, uint8_t trn, uint8_t sblk);
 void impl_uitrack_change_tn(int tn, enum topo_turnout_state v);
 void impl_uitrack_change_tn_reserv(int tn, int train);
 void impl_uitrack_change_pres(uint32_t bitfield);
